power_rangers.html
In 1993, the "Mighty Morphin’ Power Rangers" premiered on TV, fighting forces of evil. The five original rangers were each represented by a different color:

❤️‍🩹 Red
💙 Blue
🖤 Black
💛 Yellow
💖 Pink
Create a new file called power-rangers.html.

Copy/paste this code in the <body> element:

<div class="ranger-div" id="red-ranger"></div>
<div class="ranger-div" id="blue-ranger"></div>
<div class="ranger-div" id="black-ranger"></div>
<div class="ranger-div" id="yellow-ranger"></div>
<div class="ranger-div" id="pink-ranger"></div>

Insert a <style> element in the <head>. Then, apply the following styles in the <style> element:

A width of 50% and height of 100px for <div> divider elements with the ranger-div class.
A different background-color for each <div> divider element based on their id.