top_8.html
For an entire generation, a website called MySpace was the entry to the world of social networks. Tom was everyone's friend.

One of the iconic features of the MySpace era was the “Top 8", which allowed users to select eight friends to display on their profile page. It changed the social dynamics in schools everywhere:

“OMG! Did you remove so-and-so from top 8?"
“You've put me in your top friends! BFF <3."
“What's-her-name and what's-his-face are not friends?!"
Let’s create your own MySpace Top 8 with HTML!

Make a new file called top_8.html.

We'll provide you with a <style> element with the styles already inside. Go ahead and copy/paste this into the <head> of your file:

<style>
  body {
    width: 85%;
    margin: auto;
  }

  h1 {
    text-align: left;
  }

  img {
    border: 3px solid blue;
  }

  #top-8-wrapper {
    text-align: center;
  }

  .friend-card {
    display: inline-block;
    margin: 1px;
    text-align: center;
  }

  .friend-name {
    color: blue;
  }
</style>

Next, let’s add the HTML.

Add a <div> element with an id of "top-8-wrapper".
Inside, add a <h1> heading element that says "My Top Friends!", followed by two <div> elements with a class of "top-8-row".
Inside each <div> with a "top-8-row" class, add four more <div> elements that each have a class of "friend-card".
Inside each <div> with a "friend-card" class, add the following:
A <h2> header element with a "friend-name" class and the friend's name inside the tags.
An <img> image element with a src and alt attribute.
So who’s in your top 8?

If you don't want to use real names, feel free to add funny usernames or create superlatives ("class clown", "life of the party", "most likely to become president", "best hair", "most likely to be internet famous", etc).

When you are done, screenshot the web page and share it online (and don't forget to tag your top 8)!