02. Syntax.html
Let's practice writing some CSS!

Copy and paste the following HTML:

<!DOCTYPE html>
<html lang="en">
  <head>
    <link href="styles.css" rel="stylesheet">
    <title>Commentary</title>
  </head>
  <body>
    <p>Malcolm in the Middle</p>
  </body>
</html>

Next, it's time to move on to the styles.css file tab.

Select the p paragraph element and assign the following properties and values:

A color set to red.
A font-size set to 25px.
Note: Don't worry too much about what these properties and values mean or do. We just want you to get used to writing CSS rules!

Save the file and see what happens to our page. Feel free to comment this out, by surrounding the code with /* and */ to see the text revert back to its original position.